LetGO_to_Icc-server/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.0
PyJWT>=2.8
python-dotenv>=1.0
twilio>=9.0
dj-database-url>=2.0
